Introduce GfxstreamFormat
Currently, the host needs effectively 3 variables to know a format:
a GLenum "internalFormat", a FrameworkFormat, and a FeatureSet (to
know if generic YUV is NV12 or YV21).

This change introduces a new GfxstreamFormat, with the intent of this
enum class being a completely specified format, and refactors the non
frontend parts of the host (RenderControl and parts of FrameBuffer
have unfortunately become an undocumented API for other Goldfish
host components) to use the new enum class.

The motivation for this refactor is for the next change where
the vulkan compositor needs to know the swizzle of YUV formats
based on whether or not a ColorBuffer represents a UV or VU
guest gralloc buffer.

Bug: b/450007669
Test: CI
Change-Id: I2b764833c095b61757d93fb3f239e487035a9c90