Do not enable external_memoy_host mode with multiple memory types
Host allocation mode is designed for software renderers and only
supported if there is only a single type of memory. Some drivers may
still report invalid memory indices due to bugs in the extension's
implementation, but this check ensures that we can safely keep using the
memory in such cases.

Bug: 469094646
Test: CI
Change-Id: Id055234e6c9ffb3dfe7550eb03cbe1e5bc4661c2